Expose isFocused field and focus/blur subscribers
A lot happened in this commit name:
* introduces the `getElementState` method on ForageUI as a mechanism to
  query the current state of a ForageElement
* introduces the `setOnFocusEventListener` and `setOnBlurEventListener`
  methods as a mechanism to register a listener for when a focus and
  blur event occurs

The rationale is for getElementState is that apparently hasFocus() was
no reliable. During manual testing, it was not accurately reflecting
when the input was focused or not. Likely that was because we were not
adhering to the interface [that Basis Theory exposes](https://developers.basistheory.com/docs/sdks/mobile/android/events#focusevent).
Also, `isFocused` already means something in Android so it would be
against convention to hijack that name for Forage's purposes. Instead, I
took a [hint from what VGS did](https://www.verygoodsecurity.com/docs/vgs-collect/android-sdk/field-state-tracking/#:~:text=The%20getAllStates%20method%20returns%20all%20states%20of%20views%2C) which is to expose a method to query state
directly, this way avoiding hijacking of the already used `.isFocused`
field.

Other things of note in the code comments is that we needed to use
mutable pointers to the event subscribers that people will pass in order
to smooth out the differences between how BT and VGS implement their
respective event listeners