Create service to manage ForageElement state + listeners
Motivation for creating a service was to extract as much pure (business)
logic as a possible to make it easy test. Right now a lot of business
logic is wrapped up in Android Views and that is muuuuch harder to test

The ElementStateManager service is responsible for exposing a simple
API for managing the setting of event listeners for focus, blur, and
change events. It has two flavors: one for PIN and one for PANs to
reflect the subtle differences in isValid-related logic.