Create service to manage ForageElement state + listeners (#53)
* Create service to manage ForageElement state + listeners

Motivation for creating a service was to extract as much pure (business)
logic as a possible to make it easy test. Right now a lot of business
logic is wrapped up in Android Views and that is muuuuch harder to test

The ElementStateManager service is responsible for exposing a simple
API for managing the setting of event listeners for focus, blur, and
change events. It has two flavors: one for PIN and one for PANs to
reflect the subtle differences in isValid-related logic.

* Fix lint issues

* Make vault field non-null

Within the init method we use an if-else so vault is
guaranteed to point to either VGS or BT. By making `vault`
non-null we can clean up the file and make it easier to read

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Update ForageUI interface to expose event listeners

Also formally require that ForagePANEditText implements the ForageUI
interface. And, clean up the event listeners signatures of PIN to
use the new Simple/Stateful event listeners

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Wire up ForageElements to ElementStateManagers

ElementStateManagers are capable of tracking event
listeners and define the allowed state changes for
Elements. They can replace the logic that was previously
coupled the AndroidViews while consolidate the logic into
a single service that is easy to test.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Fix lint issues

* Have isValid, isEmtpy rely on state manager

Ideally we depreciate these fields entirely but, in the mean time, these
fields should be powered by the `ElementStateManager` service since that
previous facillities are no longer supported.

* Fix isValid copy/paste typo

* BREAKING: remove isValid, isEmpty from ForageUI

The recommended way of accessing observable state on
Forage Elements (e.g. PAN/PIN) is via `.getElementState()`.

We don't want to add surface-level attributes to ForageElements
because they inherit from so many other Android View's and there
is a strong chance of name collisions; promoting
`.getElementState()` gives us our own namespace and distinct
abstraction

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>