BREAKING: `tokenizeEBTCard` accept panForageEditText as arg
Previously `tokenizeEBTCard`'s implementation was relying on
the sharing of `panEntry` within the `ForageSDK` singleton
object. Sharing state is generally considered bad practice
because it does not scale to multiple parallel consumption
well.

This commit decouples (and deletes) the shared state. Now,
`tokenizeEBTCard` accepts a ForagePANEditText as an argument
and has no reliance on shared state. The end result is:
* easier to test the `shouldTokenize` logic since it can be
  extract to a service and not live in the view layer near the
  shared state
* no shared state means multiple instance of ForagePANEditText
  can co-exist without issue

Also, `getPanNumber()` was returning the raw input which is not OK now
that we have formatting. Instead, we need to return the digits only.

Signed-off-by: Devin Morgan <devin@joinforage.com>