Enrich PAN-validation with State IIN logic (#57)
* Enrich PAN-validation with State IIN logic

Opted to give `.isComplete`, `.isValid`, and `.validationError` their
own setter methods for single responsibility + readability.

Also, for the tests I was torn between create duplicate tests for each
of these fields vs having a single suite that did multiple assertions.
Basically I opted for increased extensibility at the cost of a bit more
of a maintenence burden. I thought that tradeoff was fine but maybe
other people will have opinions

* Fix linter errors

* Address Danilo's feedback

* Update isValid to be OK with too short card #

Danny flagged this in the PR but the criteria for being valid are that
the string:
1) has a legitimate StateIIN and
2) is not longer than the expected length.

The difference between validation and and completeness is that
completeness requires that the length be match the expected length of
the state.