Wire up the FormatPanTextWatcher to ForagePANEditText
For the time being, I was unable to figure out a way to support the
formatting, which uses whitespace characters, with the numbers-only
keypad that comes from having `inputType = InputType.TYPE_CLASS_NUMBER`
So, for now, supporting formatting makes the keyboard the regular
keyboard. The `FormatPanTextWatcher` removes any non-digit
non-whitespace characters.

In a follow up commit, I'll need to make sure that `ForagePANEditText`
exposes a way to return the textinput as numbers only to not break
downstream consumers of the PAN text