Wire up ForagePANEditText to support whitelisting
We'll either use the whitelist enabled validations during non-prod
environments or the strict StateIIN validtions in a prod envrionment. In
addition, we needed to remove the validation logic in the
`afterTextChanged` method because it is no longer appropriate. The
validation logic that we want to use lives in the PanElementStateManager