[FX-456] Support Whitelisted PAN Cards in Non-Prod environments (#69)
* Revert "Short circuit validation for select cards in non-prod environments (#65)"

This reverts commit e35e818a6fb5ecfc599f37297f164162af41cdaf.

* Use the plugin pattern for PanElementStateManager

Previously the `PanElementStateManager` had coupled its validation logic
to the validating against legitimtate StateIINs. However, requirements
have since been revised to allow developers to use whitelisted PANs
to invoke specific error codes during development. To accomodate, this
commit decouples the specific of validation logic to rely on pluginable
PanValidators

For the tests, it also seemed appropriate to consolidate several test
classes that were each previously only making one assertion per test
into a single class with multiple assertions per test. It seems this was
a clear maintainability win since refactoring to the pluginable
validation logic meant those three test classes were misaligned with the
newer implementation

* Wire up ForagePANEditText to support whitelisting

We'll either use the whitelist enabled validations during non-prod
environments or the strict StateIIN validtions in a prod envrionment. In
addition, we needed to remove the validation logic in the
`afterTextChanged` method because it is no longer appropriate. The
validation logic that we want to use lives in the PanElementStateManager

* Add support for whitelisted PAN to be 16 to 19 long