[FX-404] Support dynamic PAN masking based on StateIIN (#67)
* Move cardNumber StateIIN helpers to StateIIN.kt

These helpers are going to be consumed by more than just the
ElementStateManagers so putting them in a more centrally located home
to prevent random modules importing from ElementStateManager's module
would likely improve code organization

* Introduce standalone FormatPanTextWatcher service

This text watcher is responsible for:
1. formatting PAN number according to predetermined patterns for
   different PAN lengths
2. restricts allowed characters to digits for users
3. ensuring the that cursor (selector) behaves intuitively as the
   input's format evolves with the digitis entered
4. supporting pasting by filtering out non-digits and updating format

Caveats:
* For some reason, inserting characters moves the cursors position even
  though no characters are entered (only digits are allowed)
* supporting this `TextWatcher` means the input will need to accept more
  than just numbers which means we no longer display the numbers-only
  keyboard

* Wire up the FormatPanTextWatcher to ForagePANEditText

For the time being, I was unable to figure out a way to support the
formatting, which uses whitespace characters, with the numbers-only
keypad that comes from having `inputType = InputType.TYPE_CLASS_NUMBER`
So, for now, supporting formatting makes the keyboard the regular
keyboard. The `FormatPanTextWatcher` removes any non-digit
non-whitespace characters.

In a follow up commit, I'll need to make sure that `ForagePANEditText`
exposes a way to return the textinput as numbers only to not break
downstream consumers of the PAN text

* Make android.test a testImplementation dependency

* Remove 19-char limit from ForagePANEditText

This was useful preformatting but now that we format digits, the text
will include spaces that will make the inputs up to 23 digits long. Also
the character limiting (truncating) now happens in the
`FormatPanTextWatcher` so we can delete this.

* Modify ForagePANEditText to store digits only

Previously no sanitization was needed because it was a numbers-only
TextInputEditText. Now that we have formatting, the text may include
whitepsace would would mark everything as invalid and, if it did go
through, would make `payments-api` upset for unformatted PANs. This
commit ensures that only digits get passed downstream