Make sure that onChangeEventHandler gets fired once (#71)
* Make sure that onChangeEventHandler gets fired once

There are two updates to the `TextInputEditText` when the user types
something. 1) is the actual user's update and 2) is the update that
reformats the input field based on the lastest changes.

Previously, the `ForagePANEditText` was invoking the changEvent callback
for each of these, which is suboptimal. To fix this, we defer the
responsibility of knowing to invoke the user's registered callback to
the `FormatPanTextWatcher` since it holds the source of truth to
distinguish between true user events and formatting updates