Show full PaymentCapture ForageError in seperate TextView
The previous implementation showed the full ForageError in the
`snapResponse` and actually did not (I forgot) show the full ForageError
in the `cashResponse` TextView.

This commit does two things:
1) creates a seperate TextView to display the full ForageError so that
   it does not mess with the original `snapResponse` / `cashResponse`
   TextViews, which are tightly coupled to the qa-test
2) ensures that the full ForageError displays for EBT Cash payments as
   well as SNAP

Also: this PR adjusts some of the xml constraint attributes so that they
read more easily (at least for me). Now the top contstraint comes first
and the bottom contraint comes last so the hope is it reads in order
now, which is less confusing.