Add details field to ForageErrorObj (#86)
* Support details field in ForageErrorObj

All SQSMessages get parsed via
`Message.ModelMapper.from(response.data)`
this method uses
`SQSError.SQSErrorMapper.from` to parse the errors form the SQSMessage.

This commit modifies the later method and introduces a sealed class of
possible `ErrorMessageDetails`. This results in type-safe details field
in ForageErrorObj that have them. For now this is only `ebt_error_51`
but this could easily evolve over time.

* Add developer-friendly comments to ForageError

* Rename to EbtError51Details

The hope is that having the type name resemble the error code,
developers are less likely to make mistakes when type casting
and, if they do make mistakes, it should be easier to identify which
types are associated with a given "ebt_error_*"

* Show full PaymentCapture ForageError in seperate TextView

This commit does two things:
1) creates a seperate TextView to display the full ForageError so that
   it does not mess with the original `snapResponse` / `cashResponse`
   TextViews, which are tightly coupled to the qa-test
2) ensures that the full ForageError displays for EBT Cash payments as
   well as SNAP

Also: this PR adjusts some of the xml constraint attributes so that they
read more easily (at least for me). Now the top contstraint comes first
and the bottom contraint comes last so the hope is it reads in order
now, which is less confusing.

* Remove scrollview that's breaking mobile QA tests

I made this ticket to bring back in the future:

https://linear.app/joinforage/issue/FX-548/android-update-payment-capture-view-to-have-scroll-view

The problem is that this scrollview is breaking the XPATH referenced
here:
https://github.com/teamforage/mobile-qa-tests/blob/1204e26fa066d8003ba5c9487c1ffa9f0f774643/android/viewobjects/capture_payment_view.py#L23

So making this view scrollable would also require a change to the mobile
QA repo. However, I don't have that set up on my local machine and that
would take a few hours to make sure I have it working so I'll defer that
to later as technical debt / cleanup