Remove TextWatcher interface from ForagePANEditText
ForagePANEditText only implemented the TextWatcher interface as part of
the singleton flow with mutating state. Specifically, ForageSDK had a
mutable reference to the PAN and we were using the TextWatcher lifecycle
methods to udpate the shared state whenever somebody typed.

Now that we are passing the ForagePANEditText to the `tokenizeEbtCard`
method directly, we do not need the shared state and thus do not need
`ForagePANEditText` to implement the TextWatcher interace.