Swap dependence on BuildConfig with StopgapGlobalState
This does not replace sample-app's dependence on BuildConfig with
StopgapGlobalState. This largely has to do with expediency and because
merchants depend on our SDK and not our sample-app

NOTE:
1. I removed several uses of the word const since StopgapGlobalState
  is runtime now and not build time
2. StopgapGlobalState will only be instantiated if `.setForageConfig`
  was called on either ForagePANEditText or ForagePINEditText. If it
  was not called, there will be runtime exceptions!!
3. I would not be surprised if for qa-tests or for more rigorous
testing, it becomes important to replace sample-app's dependence on
BuildConfig with it's own version of global state. HOWEVER, it cannot depend
on StopgapGlobalState and it would need to depend on its own instance of global
state: they are different packages after all!

Signed-off-by: Devin Morgan <devin@joinforage.com>