Dropped both ForageElements dependence on BuildGradle
Getting ForagePANEditText and ForagePINEditText to not depend on
BuildConfig and instead depend on StopgapGlobalState was muuuuch
more involved compared to the previous where the effort was largely
find and replace.

Swapping the dependence in Forage***EditText required teasing out the
parts of the `init { ... }` method that depended on BuildConfig in a way
that still cooperated with the render lifecycle of Android Views.

The basic strategy employed for both Forage***EditText was moving all of
the decisions that depend on knowing the ForageConfig (i.e. the environment)
to the `setForageConfig` method. And, instantiate both all other possibl
views.

Essentially, `init { ... }` would hedge to support all possible decision
outcomes in cases where it was not trivial to just punt the decision
entirely.

Signed-off-by: Devin Morgan <devin@joinforage.com>