Get rid of MissingForageConfigException
As I was writing up the migration guide, I realized that there was an
opportunity to keep things simple by throwing a single type of exception
and varying the message. This would reduce confusion about what the
difference between `MissingForageConfigException` and
`ForageConfigNotSetException` are. In fact they refer to the same issue
but that issue can happen at two different points depending on how the
merchant implements the Forage Android SDK