Remove TextWatcher interface from ForagePANEditText (#93)
* Remove TextWatcher interface from ForagePANEditText

ForagePANEditText only implemented the TextWatcher interface as part of
the singleton flow with mutating state. Specifically, ForageSDK had a
mutable reference to the PAN and we were using the TextWatcher lifecycle
methods to udpate the shared state whenever somebody typed.

Now that we are passing the ForagePANEditText to the `tokenizeEbtCard`
method directly, we do not need the shared state and thus do not need
`ForagePANEditText` to implement the TextWatcher interace.

* [BREAKING] Milestone 1 - Add Proposed Android v3 Interfaces (#97)

* Add proposed new interfaces for Android v3

As dicussused
[here](https://www.notion.so/joinforage/Android-3-0-Design-Proposal-1ee6be95544d4dc78e7a7e7c88f46bbc?pvs=4)

This represents the [Milestone 1](https://www.notion.so/joinforage/Plan-for-On-Time-Delivery-of-Android-3-0-Breaking-Changes-226d8ee6f8294d2694b1bb451791960b?pvs=4#dcbc763046cb4875a3e08cf2703cfaf3) of The Plan to Deliver to GoPuff On-Time

* Update sample app to use Android v3 interface.

FYI - I previously introduced a bug in the QA-tests by adding reusable
to `false`. But I fixed it by setting to true

* apply spotless

* Rename InternalForageElement -> AbstractForageElement

I think this naming is more appropriate. It's not that it is an internal
element and its more that its an element that cannot be instantiating
but can be subclassed.

* Rename ForageContext -> ForageConfig

* Remove resolved comment

This was resolved in [this GitHub thread](https://github.com/teamforage/forage-android-sdk/pull/97#discussion_r1331707671)

* Introduce MissingForageConfigException + helpful docstrings

* apply spotless + fix typo

* Fix ForageElements typo

* Prefer foragePin and foragePan for varible names

* Milestone 2 - Introduce `StopgapGlobalState` (#102)

* Introduce EnvConfig and EnvOption

In order to migrate away from FLAVORS, we need to move all of the
constants that are currently declared in build.gradle > productFlavors
into runtime code. That is the motivation of these two additions: they
serve as convinient ways to get at config constants if you have access
to either ForageConfig or the sessionToken

* Introduce the notorious STOPGAP_GLOBAL_STATE singleton

This is a temporary hack to ship Android v3 changes in a timely manner
and defering some internal cleanup work to fully move away from
singletons. See this doc for more details:

https://www.notion.so/joinforage/226d8ee6f8294d2694b1bb451791960b

* Mark Env* as internal + fix baseUrl typos

* Address spotless lint errors

* Fix SANDBOXbox typo for vgsVaultType

Signed-off-by: Devin Morgan <devin@joinforage.com>

* rename things from ForagContext -> ForageConfig

* rename ForageContext -> ForageConfig

* Milestone 3 - Replace BuildConfig.FLAVOR with Stopgap FLAVOR (#103)

* Associate EnvOption enums with their string values

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Add static inProd method to EnvConfig

The goal support a concise way or checking if we're in prod.
Previously it would have been

```
// previous
if (EnvConfig(forageContext).FLAVOR.value == "prod")

// new
if (EnvConfigs(forageContext).inProd())
```

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Swap dependence on BuildConfig with StopgapGlobalState

This does not replace sample-app's dependence on BuildConfig with
StopgapGlobalState. This largely has to do with expediency and because
merchants depend on our SDK and not our sample-app

NOTE:
1. I removed several uses of the word const since StopgapGlobalState
  is runtime now and not build time
2. StopgapGlobalState will only be instantiated if `.setForageConfig`
  was called on either ForagePANEditText or ForagePINEditText. If it
  was not called, there will be runtime exceptions!!
3. I would not be surprised if for qa-tests or for more rigorous
testing, it becomes important to replace sample-app's dependence on
BuildConfig with it's own version of global state. HOWEVER, it cannot depend
on StopgapGlobalState and it would need to depend on its own instance of global
state: they are different packages after all!

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Dropped both ForageElements dependence on BuildGradle

Getting ForagePANEditText and ForagePINEditText to not depend on
BuildConfig and instead depend on StopgapGlobalState was muuuuch
more involved compared to the previous where the effort was largely
find and replace.

Swapping the dependence in Forage***EditText required teasing out the
parts of the `init { ... }` method that depended on BuildConfig in a way
that still cooperated with the render lifecycle of Android Views.

The basic strategy employed for both Forage***EditText was moving all of
the decisions that depend on knowing the ForageConfig (i.e. the environment)
to the `setForageConfig` method. And, instantiate both all other possibl
views.

Essentially, `init { ... }` would hedge to support all possible decision
outcomes in cases where it was not trivial to just punt the decision
entirely.

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Drop all usage of flavor variables in build.gradle

* apply spotless

* rename ForageContext -> ForageConfig

* Introduce ForageConfigNotSetException

and get rid of the outstanding TODO

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Add comments around _SET_ONLY field abstractions

* Add comment about hedging VaultWrappers

* apply spotless

* Make TokenizeEBTCardParams.reusable optional

* Get rid of MissingForageConfigException

As I was writing up the migration guide, I realized that there was an
opportunity to keep things simple by throwing a single type of exception
and varying the message. This would reduce confusion about what the
difference between `MissingForageConfigException` and
`ForageConfigNotSetException` are. In fact they refer to the same issue
but that issue can happen at two different points depending on how the
merchant implements the Forage Android SDK

* Fix possible typo: panBoxStrokeColor on PIN element styles

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Update Catalog to work with SDK v3

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Update the README.md to be in sync with V3 changes (#105)

* Update the README.md to be in sync with V3 changes

This served as a helpful exercise for flushing out details for the V3
Migration guide.

Relevant points to note:
1. Organized the Headers so that they followed a similar story flow
   while keeping the doc DRY by linking to other sections
2. I update the styling snippets to articulate all of the XML styling
   attributes that `ForagePANEditText` and `ForagePINEditText`
   respectively support
3. All submit flows now include example snippets of the View and the
   ViewModel and shows the shell of classes to make it easier for
   developers to feel oriented in the examples. Previously the snippets
   were just of the ViewModel which is only half the picture

* Rename ForageContext -> ForageConfig

* Eliminate reference to Context in submit methods

* Fix typo in TokenizeEBTCardParams definition

* Fix wrong param data class for submit methods

* Resolve outstanding TODO to changelog link

* Micellaneous fixes that Danilo caught

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>