Ensure init operations on setForageConfig happen 1x
The singleton objects for Launch Darkly and Datadog seem to already take
defensive measures to ensure they are not getting initalized more than
once. However, `setForageConfig` was also adding child views to the main
on each call instead of thoughtfully making sure that the views were
only getting add on the first call.

This comment streamlines this orchestration into the
AbstractForageElement class and renames the side-effectful operations
happening in the PIN and PAN elements to initWithForageConfig to better
articulate its intention