[FX-593] Convert floats to dimensions for styling (#109)
* Put <item> tags under correct <style> param

The tags
```xml
 <item name="panBoxStrokeWidth">5dp</item>
 <item name="panBoxStrokeColor">@color/mtrl_textinput_default_box_stroke_color</item>
```

are used to style the PAN's EditText which has the name
`TokenizeForagePANEditTextStyle` and not
`TokenizeForageTextInputLayoutStyle`. The latter is to style the
LinearLayout that houses the EditText

* Adjust panBoxStrokeWidth[Focused] to be dimensions

Previously these values were floats but they need to be dimensions so
that they can be response to different screen sizes. Floats will not be
responsive.

Also I renamed the variable:
panFocusedBoxStrokeWidth --> panBoxStrokeWidthFocused
to keep names aligned and reduce confusion

* Make stroke widths easier to view in catalog