Drop product flavors from forage-android
NOTE: A follow-up commit will drop product flavors for sample-app as
this commit does not touch sample-app

In the past (see 7efb39), we published two build artifacts:
1. the executable that could run `forage-android` / `sample-app`
2. the source code of `forage-android` / `sample-app`

As of 0aac9d, we stopped publishing the `sample-app` executable but
continued to publish `sample-app`'s source code via `androidSourcesJar`.

For completeness, this commit brings back the block of code in
`publish-module.gradle` that exposes the `sample-app` executable. That
is, I feel like we should either expose both the source code and the
executable for the `sample-app` or neither because we only want to
expose the source code and executable of `forage-android`