Drop CreatePaymentViewModel dependence on BuildConfig
Specifically, `FlowCreatePaymentViewModel.kt` relied on BuildConfig
indirectly through PaymentsRepository because PaymentsRepository needed
to know which API server to use for creating the payment.

Breaking this dependency required refactor the network module to stop
using dependency injection.

I'll note that it's unclear to me why we needed DI here in the first
place. But even if DI served a purpose, it is certainly no longer
needed now that we're breaking our dependence on BuildConfig.FLAVOR
in the sample-app.

This commit removes all usage of DI related to PaymentsRepository's
usage. Further, breaking the dependence on BuildConfig.FLAVOR motivated
the introduction of the EncConfig object which is a dumbed down version
of EnvConfig from the `forage-android` module.