Stop using DI for Moshi in ViewModels
While DI has a role, in this context it heavily obfuscates the flow of
data especially for a team of relatively new android developers such as
myself.

In the interest of expediency, I did not tease out the copy/pasted usage
of
```
val moshi = Moshi.Builder()
   .addAdapter(Rfc3339DateJsonAdapter().nullSafe())
   .build()
```
into its own method. We can certainly do that later though.