Drop TokensViewModel's dependence on BuildConfig
In a typical app that is consuming ForageSDK, that app would have its
own way of declaring the environment it is operating within.

In `sample-app`, however, we rely on inferring the environment from the
session token that a developer passes into the app (either by
painstakingly entering it into the emulator or by copy/pasting it into
the source code before building).

Because the session token determines the environment,
`TokensUIDefaultState.kt` is no longer useful, allowing us to delete it
and drop our dependence on BuildConfig