[FX-597] Drop usage of Build Flavors in Forage Android (#113)
* Remove publishVersion in gradle.properties

This was not getting consumed anywhere. It was legacy from when we used
BuildCongfig.VERSION. This is basically cleanup work

* Drop usage of BuildConfig.VERSION

Augments the EnvConfig object to expose a PUBLISH_VERSION variable,
which represents same value as PUBLISH_VERSION declared in the
gradle.build file

* Drop product flavors from forage-android

NOTE: A follow-up commit will drop product flavors for sample-app as
this commit does not touch sample-app

In the past (see 7efb39), we published two build artifacts:
1. the executable that could run `forage-android` / `sample-app`
2. the source code of `forage-android` / `sample-app`

As of 0aac9d, we stopped publishing the `sample-app` executable but
continued to publish `sample-app`'s source code via `androidSourcesJar`.

For completeness, this commit brings back the block of code in
`publish-module.gradle` that exposes the `sample-app` executable. That
is, I feel like we should either expose both the source code and the
executable for the `sample-app` or neither because we only want to
expose the source code and executable of `forage-android`

* Drop CreatePaymentViewModel dependence on BuildConfig (#114)

* Drop CreatePaymentViewModel dependence on BuildConfig

Specifically, `FlowCreatePaymentViewModel.kt` relied on BuildConfig
indirectly through PaymentsRepository because PaymentsRepository needed
to know which API server to use for creating the payment.

Breaking this dependency required refactor the network module to stop
using dependency injection.

I'll note that it's unclear to me why we needed DI here in the first
place. But even if DI served a purpose, it is certainly no longer
needed now that we're breaking our dependence on BuildConfig.FLAVOR
in the sample-app.

This commit removes all usage of DI related to PaymentsRepository's
usage. Further, breaking the dependence on BuildConfig.FLAVOR motivated
the introduction of the EncConfig object which is a dumbed down version
of EnvConfig from the `forage-android` module.

* Drop TokensViewModel's dependence on BuildConfig

In a typical app that is consuming ForageSDK, that app would have its
own way of declaring the environment it is operating within.

In `sample-app`, however, we rely on inferring the environment from the
session token that a developer passes into the app (either by
painstakingly entering it into the emulator or by copy/pasting it into
the source code before building).

Because the session token determines the environment,
`TokensUIDefaultState.kt` is no longer useful, allowing us to delete it
and drop our dependence on BuildConfig

* Formally stop using build flavors in sample-app

In previous commits we dropped the sample-apps various dependencies on
BuildConfig. With no more dependencies on BuildConfig, we can not drop
build flavors entirely from the `sample-app`'s `build.gradle`

* Delete a random unused file

* Stop using DI for Moshi in ViewModels

While DI has a role, in this context it heavily obfuscates the flow of
data especially for a team of relatively new android developers such as
myself.

In the interest of expediency, I did not tease out the copy/pasted usage
of
```
val moshi = Moshi.Builder()
   .addAdapter(Rfc3339DateJsonAdapter().nullSafe())
   .build()
```
into its own method. We can certainly do that later though.

* Testing changes to CI workflow in seperate file (#115)

* Modify CI.yaml to not rely on build flavors

It turns out that when you modify a github actions workflow file, the
changes are reflected immediately in github. So, if you investigate the
CI job run on this PR, you'll see that it's already using the latest
changes from this commit!

A few remarks on the code changes:
* I assumed we want to use debug for everything. Previously we were
  suing sandbox for the unit tests and staging for the Appium tests.
  However, since we have dropped flavors, `forage-android` will infer
  the envionrment based on the sessionToken and the mobile-qa-tests
  [already rely on
  staging](https://github.com/teamforage/mobile-qa-tests/blob/7f15c11934c8211a38baff87a30ee69b04191031/helpers/forage_api.py#L8C12-L9)
  so it technically makes no difference at present as to whether we
  choose debug vs release. But, in case something changes in the future,
  I figure it was safer to error on using debug for everthing

* Modify Lint.yml to rely on dbug and not cert