Use datadog v2 in Log.kt
In general, we want to use the latest version of datadog so that we can
have the latest features and security fixes. Previously we thought we
needed v2 of datadog because of the singleton design pattern used in v1.
However, that would not matter now that we have shadowed the datadog lib
into forage as the forage-namespaced version uses completely different
memory