Create gradle plugin that shadows AAR artifacts
Datadog is an AAR and not a JAR. The [Shadow Gradle
Plugin](https://github.com/johnrengelman/shadow) is capable of
manipulating JAR artificats so that their contents are under a new
namespace of our choosing, which is what we want to do.

The ShadowAarDependenciesPlugin takes some information about a generic
AAR artifact and will extract the classes.jar artifact and then apply
the Gradle Shadow Plugin.

The end result for Forage is we get to use Datadog effectively the same
way as before except now we do:

// new
import com.joinforage.datadog...

// old
import com.datadog...