Add datadog v1 to sample app
This will be a passive way for us to ensure that including datadog v2 in
forage-android passes the build step. As long as we use our custom
gradle plugin, including datadog v1 in the sample-app should not ber a
problem. If something ever stops working in the gradle plugin, this will
fail to build (i.e. make noise)