Use datadog v2 in Log.kt (#122)
* Use datadog v2 in Log.kt

In general, we want to use the latest version of datadog so that we can
have the latest features and security fixes. Previously we thought we
needed v2 of datadog because of the singleton design pattern used in v1.
However, that would not matter now that we have shadowed the datadog lib
into forage as the forage-namespaced version uses completely different
memory

* Update several core packages to newer versions

This will be necessary for the shadow-plugin for datadog to work

Signed-off-by: Devin Morgan <devin@joinforage.com>

* Create gradle plugin that shadows AAR artifacts

Datadog is an AAR and not a JAR. The [Shadow Gradle
Plugin](https://github.com/johnrengelman/shadow) is capable of
manipulating JAR artificats so that their contents are under a new
namespace of our choosing, which is what we want to do.

The ShadowAarDependenciesPlugin takes some information about a generic
AAR artifact and will extract the classes.jar artifact and then apply
the Gradle Shadow Plugin.

The end result for Forage is we get to use Datadog effectively the same
way as before except now we do:

// new
import com.joinforage.datadog...

// old
import com.datadog...

* Have CI use JDK 17 instead of 11

If this doesn't pan out we can revert. This doesn't change the target
version that the APK get's compiled down to

* Add datadog v1 to sample app

This will be a passive way for us to ensure that including datadog v2 in
forage-android passes the build step. As long as we use our custom
gradle plugin, including datadog v1 in the sample-app should not ber a
problem. If something ever stops working in the gradle plugin, this will
fail to build (i.e. make noise)

* Update readme

We want to document that people need to use Android Studio Giraffe
or above. And also that you need to use JDK 17

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>