Remove LDManager's dependence on StopgapGlobalState
LDManager relied on StopgapGlobalState for the ld client key. The fix
was to pass the ld client key to LDManager via the initialize method.

There was an opportunity to reduce the complexity of LDManager by
exposing seperate mechanisms for building test-related configs and
prod-related configs.

I also took the opportunity to tease out the logic around consuming the
primaryTrafficFlag and converting it to vault provider and make that a
standalone pure function. The goal was to have getVaultProvider be able
to gaurantee that it would return a non-null value and avoid type
casting. Creating a function that maps Float -> VaultProvider allows us
to store only the Float as state on LDManager and we can gaurantee a
default value of ALWAYS_VGS