[FX-677] Remove ForageConstants dependence on StopgapGlobalState (#126)
* Remove ForageConstants dependence on StopgapGlobalState

This was accomplished by elimintating the BASE_URL constant and the
provideHttpUrl method from ForageConstants and having all ForageSDK
method infer the baseUrl which is inferred from the sessionToken

* Add tests to surface any typos in Env baseUrls

We currently build up URL objects from strings. However, these objects
need to handle the case that somebody passes a mal-formed URL so they
either say they're return a possibly-null object or that they'll capable
of throwing during runtime. Because we're unable to gain confidence that
our URls are wellformed from compile-checking, we can use these unit
tests as an alternative