[FX-681] Remove LDManager's dependence on StopgapGlobalState (#128)
* Remove LDManager's dependence on StopgapGlobalState

LDManager relied on StopgapGlobalState for the ld client key. The fix
was to pass the ld client key to LDManager via the initialize method.

There was an opportunity to reduce the complexity of LDManager by
exposing seperate mechanisms for building test-related configs and
prod-related configs.

I also took the opportunity to tease out the logic around consuming the
primaryTrafficFlag and converting it to vault provider and make that a
standalone pure function. The goal was to have getVaultProvider be able
to gaurantee that it would return a non-null value and avoid type
casting. Creating a function that maps Float -> VaultProvider allows us
to store only the Float as state on LDManager and we can gaurantee a
default value of ALWAYS_VGS

* Stop caching LDManager.getVaultProvider

After speaking with both Danny and Nitin, it emerged that we no longer
need to cache the value returned by getVaultProvider. This updated
criteria allowed us to remove all notion of TEST_ only functionality on
LDManager.

Other updates in this commit:
* remove logger state from LDManager singleton because it can just as
  easily be passed to each function. There's no real win from caching
  the Logger on the LDManager singleton and that also would prevent
  future opportunities for using a Logger pattern where Loggers can
  return new loggers that share additional context
* combine the testing of the default value of getVaultProvider of VGS
  with the test that getVaultProvider honors feature flag changes. This
  was a quick and dirty thing to get around annoying sharing of state
  across test issues I was facing