Update unit tests for usState support
Had to do a bit of type casting in the unit tests to get around the
`Nothing?` type of PinDetails giving the compiler ambiguous overloading
but that wasn't too big an issue.

It's worth noting that making ElementState generic did break the tests
(i.e. it was not reverse compatibile) because the tests explicitly
referenced the types `ElementState<PanDetails>` and
`StatefulElementListner<PanDetails>` callback. Referencing these types
in this way is not necessary and it's like that GoPuff did not do this
in which case it would not even be a breaking change for them.