[FX-722] Add support for `derivedCardInfo.usState`  to `ElementState` for `ForagePANEditText` (#133)
ElementState previously only cared about values that were common to both
PAN end PIN. We have reached a point where the states will need to
diverage.

* Update unit tests for usState support

* Make ElementState an interface instead of data class

However, Forage X had a conversation around this where we said we would
be OK with risk of changing the type of `ElementState` to something
other than `data class`. This opened the door for us to make
`ElementState` an interface, which is decoupled from the actual Data
Transfer Objects used to ferry the state of the ForageElement.

The rest of the changes in this commit follow from making `ElementState`
an interface. One upside of this approach is that we get rid of
`PinDetails`, which was annoying