Update sample app to show usState of PAN (#134)
* Update sample app to show usState of PAN

The change of adding `details` to `ElementState` by making
`ElementState` generic appears to be non-breaking becasue the compiler
did not complain before making this commit (i.e. before modify the
sample app).

* Have launch darkly block main thread for up to 1s

We're noticing in production that GoPuff Android apps are not respecting
dynamic feature flag values. We played around with this in dev and it
appeared that LDClient was reading the default VGS flag becuase we were
not giving it any time to fetch the latest feature flag values.

We're trying to be a bit descrite with this change because we don't
think it will cause any issues and we think GoPuff may overreact a bit
if they learn we're allowing Forage to block the main thread