Add the CPay SDK JAR as a library dep in Sample App
A lot of subsequent work is going to be leveraging the CPay SDK inside
the sample app. This is a simple commit to lay the foundation for that.

It's possible that in Android Studio, in order to get type support for
the CPay SDK JAR, that you need to explicitly add the JAR as a library

See here: https://stackoverflow.com/a/35241990/10541855