Move terminalId to POSComposeApp (#172)
* Move terminalId to POSComposeApp

`terminalId` is similar to `sessionToken` and `merchantId` in that
it likely only gets set once with the initial fetch of merchant
details and then, afterwards, will not need to be set again. It
could follow from this that `terminalId` should live alongside
`merchantId` and `sessionToken` inside `POSUIState`.

I am not opposed to `terminalId` ultimately living in
`POSUIState`. However, reading the `terminalId` from the K9SDK
is an asynchronous process because we need to wait for the
K9SDK to finish initializing before we can read the `terminalId`
value.

It was not readily apparent to me how we would request the K9SDK
to initialize once and only once from within `POSComposeApp` and
then after the async initialization has completed, read the value
of `terminalId`. I do believe it's possible to do this and once we
figure this out, this would probably be the better way to go.

In the interest of speed, I opted for using `remember` and then
reading the value of `k9SDK.terminalId` as needed throughout
`POSComposeApp`

---------

Signed-off-by: Devin Morgan <devin@joinforage.com>