Expose fetched `merchant` obj conveniently
We ideally do not need to write this if-statement everywhere that
we want to access the fetched merchant value (or `null` if the
merchant was not fetched).

We can solve this problem by either modifying the sealed
`MerchantDetailsState` interface/class or by adding an accessor
like the one added in this commit.

The convenient accessor appears good enough for now but we can
revist in the future as our needs change

Signed-off-by: Devin Morgan <devin@joinforage.com>