Revert pos-changes made to online-only data models
So it turns out that PaymentsAPI is sending different shaped
responses for POS vs Online-Only flows. And given that we're
using retrofit to generate the response data accessors, we'll
start getting runtime errors if we continue to share the data
models. This is because fields that may exist or may not be
null for one could be the opposite in the other.

This commit takes the first step of decoupling shared data
models by returning the online-only data models back to
their original schemas

Signed-off-by: Devin Morgan <devin@joinforage.com>