Add optional `.init` to ForageTerminalSDK
Right now we are pretty sure that the boot-up process involving
initializing a all the necessary encryption keys in the
`ForageTerminalSDK` will take up to 10 seconds. That is probably too
long for Forage to insist on lazily booting up the ForageTerminalSDK
when any submit methods are called.

So, we add the `.init` method as a mechanism for Toast or other
merchants to be able to run the longer-running init process ahead of
time so that submit methods are snappier.

By making `.init()` optional, we leave open the opportunity to
depreicate `.init()` in the future if we find out that the startup
processes are faster than we feared.