Create / organize receipt layout service
The original `pos/receipts` service defined three
level of abstraction for expressing the layout of a Receipt:
1) `ReceiptLinePart` - smallest unit of text in a layout. Akin
   to  a cell in table
2) `ReceiptLayoutLine` - consists of multiple
   `ReceiptLinePart`s. Akin to a row in a table
3) `ReceiptLayout` - consists of multiple `ReceiptLinePart`s
   and represents an entire receipt. This is the top-most
   level of the abstraction and can be physically printed or
   displayed in an Android View

These three models made it possible to represent receipts in
a common way. What still needed to be done - and what this
commit does - is define a convenient way of mapping data from
the Payments API into these receipt building blocks.

This commit defines a class for each test scenario in
certification. The class accepts constructor params that are
friendly towards the data that Payments API returns. Each class
also exposes a method for expressing this data as a
ReceiptLayout which can be easily viewed or printed.

This commit attempts to reduce duplication by
acknowledging that all test scenario receipts share
 a common template.

Signed-off-by: Devin Morgan <devin@joinforage.com>